icu_properties: expose Numeric_Value property
a lot of the generation code for this is very similar to what was done
for the bidi mirroring glyph property, as both of these are the same type
of property: an [[enum_property]], but without discriminants.

the concrete values provided in the nv.toml file are encoded in a very
specific matter to make them losslessly fit into an i32, and therefore
need to be decoded before you can actually do something with them.
for now (similar to the icu4c library), this only provides a way to get
the unencoded value as an f64.

in the future i might open another pull request to give access to the
numeric value losslessly as some sort of Fraction struct, but for now
this should be good enough for the initial implementation.